fix pathing for main
